+++ src/test/scala/org/podval/xml/XmlWriterSpec.scala
@@ -1,6 +1,5 @@
 package org.podval.xml
 
-import Xml.given
 import org.scalatest.funsuite.AnyFunSuite
 
 final class XmlWriterSpec extends AnyFunSuite: